stocks-app-py-2018-master/requirements.txt
python-dotenv
ipython
requests
pytest
csv
json
os
datetime
